requirements.txt
bcrypt==3.2.2
blueprint==3.4.2
cffi==1.15.0
click==8.0.4
colorama==0.4.4
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.4
itsdangerous==2.1.1
Jinja2==3.0.3
MarkupSafe==2.1.0
Pillow==9.1.1
pycodestyle==2.7.0
pycparser==2.21
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.37
toml==0.10.2
turtle==0.0.1
Werkzeug==2.1.2
WTForms==2.1
zipp==3.8.0
